requirements.txt
asgiref==3.8.1
boto3==1.34.14
botocore==1.34.162
certifi==2025.1.31
Django==4.2.9
django-cors-headers==4.3.1
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jmespath==1.0.1
mysqlclient==2.2.7
packaging==24.2
pip-system-certs==4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wrapt==1.17.2
whitenoise==6.6.0
cryptography==43.0.3
requests==2.32.3
